set new startView identifier Doctor Integer -> Long

+++ src/main/java/de/adesso/termacare/database/repo/DoctorRepo.java
@@ -3,7 +3,7 @@ package de.adesso.termacare.database.repo;
 import de.adesso.termacare.database.construct.AbstractRepo;
 import de.adesso.termacare.data.entity.Doctor;
 
-public class DoctorRepo extends AbstractRepo<Doctor, Integer>{
+public class DoctorRepo extends AbstractRepo<Doctor, Long>{
 
 	public DoctorRepo(){
 		super(Doctor.class);
